requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.13.0
numpy==1.24.3
pillow
h5py